Updated code for asynchronous page loading
This is a required update for plug-ins by April 1. I also just pushed this live to Figma community.

Figma's migration doc says to set up this complex Typescript and linting dev environment. I was not able to get it to work, so I just ignored it and don't use linting. https://www.figma.com/plugin-docs/migrating-to-dynamic-loading/